// sanity.types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};

export type CategoryReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "category";
};

export type BrandReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "brand";
};

export type Product = {
  _id: string;
  _type: "product";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  slug?: Slug;
  image?: Array<{
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
    _key: string;
  }>;
  description?: string;
  price?: number;
  discountPrice?: number;
  categories?: Array<
    {
      _key: string;
    } & CategoryReference
  >;
  stock?: number;
  brand?: BrandReference;
  status?: "new" | "hot" | "sale";
  variant?: "power" | "hand" | "tech" | "fasteners" | "other";
  isfeatured?: boolean;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};